adopt the new api, split everything into small files
