+++ examples/bindings.ts
@@ -1,20 +1,20 @@
-import { shorthand } from '../src/index'
+import { vnodeHTML } from '../src/index'
 
-export const a = shorthand('a')
-export const article = shorthand('article')
-export const button = shorthand('button')
-export const div = shorthand('div')
-export const footer = shorthand('footer')
-export const h1 = shorthand('h1')
-export const h3 = shorthand('h3')
-export const h4 = shorthand('h4')
-export const header = shorthand('header')
-export const img = shorthand('img')
-export const input = shorthand('input')
-export const label = shorthand('label')
-export const li = shorthand('li')
-export const p = shorthand('p')
-export const span = shorthand('span')
-export const section = shorthand('section')
-export const strong = shorthand('strong')
-export const ul = shorthand('ul')
+export const a = vnodeHTML('a')
+export const article = vnodeHTML('article')
+export const button = vnodeHTML('button')
+export const div = vnodeHTML('div')
+export const footer = vnodeHTML('footer')
+export const h1 = vnodeHTML('h1')
+export const h3 = vnodeHTML('h3')
+export const h4 = vnodeHTML('h4')
+export const header = vnodeHTML('header')
+export const img = vnodeHTML('img')
+export const input = vnodeHTML('input')
+export const label = vnodeHTML('label')
+export const li = vnodeHTML('li')
+export const p = vnodeHTML('p')
+export const span = vnodeHTML('span')
+export const section = vnodeHTML('section')
+export const strong = vnodeHTML('strong')
+export const ul = vnodeHTML('ul')
