+++ src/index.ts
@@ -1,3 +1,6 @@
-export * from './core'
+export * from './vnode'
+export * from './dom'
+export * from './render'
+export * from './utils'
 export * from './state'
 export * from './webc'
