+++ src/utils.ts
@@ -1,0 +1,5 @@
+export const vnodeHTML = <TName extends keyof HTMLElementTagNameMap>(elName: TName) =>
+  vnode<HTMLElementTagNameMap[TName], HTMLElementEventMap>(() => document.createElement(elName))
+
+export const frag = vnode(() => document.createDocumentFragment())
+export const styleEl = vnodeHTML('style')
